set working prototype
We removed generateNextMoves() because its current implementation is buggy.  However, you can now move the knight to any position and it will generate its next possible moves from those positions, but not recursively. We will save that for a later implementation via Breadth-First Searching

This will be helpful later on.